move getDebugConfigFolder to utils
